change the way of handling http requests in single task, naming, vars types

+++ source/picture_management/picture_logic_utils.c
@@ -2,7 +2,7 @@
 
 #include <assert.h>
 
-uint8_t getLastPictureIndex( const PictureDataView_t *picture )
+int32_t getLastPictureIndex( const PictureDataView_t *picture )
 {
     assert( picture );
